requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.6.1
pydantic-settings==2.2.1
asyncpg==0.29.0
httpx>=0.24.0
python-telegram-bot==20.7
apscheduler==3.10.4
python-dotenv==1.0.1
